requirements.txt
feedparser
google-generativeai>=0.8.3
duckduckgo-search>=6.0.0
trafilatura